ver0.6 sonarcloud rev (#225)
* sonar code rev

* code comment

* fix linux build take 1

* fix linux build take 2

* sonarcloud rev

+++ sdl2-sonic-drivers/src/hardware/opl/woody/WoodyOPL.cpp
@@ -41,7 +41,7 @@ namespace hardware
             {
                 init();
             }
-            void WoodyOPL::write(const int port, const int val) noexcept
+            void WoodyOPL::write(const uint32_t port, const uint8_t val) noexcept
             {
                 //opl->write(a, v);
             }
